feat(sdk): split Token into Token + WrappedToken hierarchy [SDK-96]
Restructures the high-level token API to mirror the ERC-7984 contract
architecture: `Token` for the base confidential token, `WrappedToken`
extending it with wrapper-specific operations.

- `Token`: read + base writes (balance, transfer, set/isOperator,
  decryptBalanceAs, batchBalancesOf, batchDecryptBalancesAs)
- `WrappedToken extends Token`: shield, unshield, unshieldAll,
  resumeUnshield, unwrap, unwrapAll, finalizeUnwrap, allowance,
  approveUnderlying, underlying

`ReadonlyToken` is removed; reads live on `Token`. `createToken(addr)`
takes a single address (no wrapper param); new `createWrappedToken(addr)`
returns a `WrappedToken` — the wrapper IS the confidential token.

Mutation factories (shield, unshield, unwrap, finalize, approveUnderlying,
resumeUnshield) now take a `WrappedToken`. `delegateDecryption` /
`revokeDelegation` mutation factories now take `(sdk, contractAddress)`,
matching the SDK-level primitives added in SDK-95.

Tests updated in-place; `delegation.test.ts` and `readonly-token.test.ts`
are removed (coverage migrated to `zama-sdk.test.ts`, `token.test.ts`,
`batch-decrypt-as.test.ts`); new `wrapped-token.test.ts` covers the
wrapper-specific surface.

Migration: `ReadonlyToken`, `createReadonlyToken`, and the two-argument
`createToken(addr, wrapper)` are removed. Callers that need wrapper
operations should use `createWrappedToken(wrapperAddr)`.
