+++ packages/sdk/src/index.ts
@@ -78,17 +78,16 @@ export { WrappersRegistry, DefaultRegistryAddresses } from "./wrappers-registry"
 export type { WrappersRegistryConfig, ListPairsOptions } from "./wrappers-registry";
 export {
   Token,
-  ReadonlyToken,
+  WrappedToken,
   type BatchBalancesResult,
   type BatchDecryptAsOptions,
-  ZERO_HANDLE,
-  isZeroHandle,
   savePendingUnshield,
   loadPendingUnshield,
   loadPendingUnshieldRequest,
   clearPendingUnshield,
   type PendingUnshieldRequest,
 } from "./token";
+export { ZERO_HANDLE, isZeroHandle } from "./utils/handles";
 export {
   MemoryStorage,
   memoryStorage,
