+++ packages/sdk/src/query/__tests__/delegate-decryption.test.ts
@@ -1,30 +1,33 @@
-import { describe, expect, test } from "../../test-fixtures";
+import { describe, expect, test, vi } from "../../test-fixtures";
 import { delegateDecryptionMutationOptions } from "../delegate-decryption";
+import type { Address } from "viem";
+
+const TOKEN = "0xaAaAaAaaAaAaAaaAaAAAAAAAAaaaAaAaAaaAaaAa" as Address;
+const DELEGATE = "0x2b2B2B2b2B2b2B2b2B2b2b2b2B2B2b2b2B2b2B2B" as Address;
 
 describe("delegateDecryptionMutationOptions", () => {
-  test("delegates delegateDecryption", async ({ mockToken }) => {
-    const options = delegateDecryptionMutationOptions(mockToken, mockToken.address);
+  test("delegates delegateDecryption", async ({ sdk }) => {
+    const spy = vi.spyOn(sdk, "delegateDecryption").mockResolvedValue({ txHash: "0x" });
+    const options = delegateDecryptionMutationOptions(sdk, TOKEN);
 
-    expect(options.mutationKey).toEqual(["zama.delegateDecryption", mockToken.address]);
-    await options.mutationFn({
-      delegateAddress: "0x2b2B2B2b2B2b2B2b2B2b2b2b2B2B2b2b2B2b2B2B",
-    });
-    expect(mockToken.delegateDecryption).toHaveBeenCalledWith({
-      delegateAddress: "0x2b2B2B2b2B2b2B2b2B2b2b2b2B2B2b2b2B2b2B2B",
+    expect(options.mutationKey).toEqual(["zama.delegateDecryption", TOKEN]);
+    await options.mutationFn({ delegateAddress: DELEGATE });
+    expect(spy).toHaveBeenCalledWith({
+      contractAddress: TOKEN,
+      delegateAddress: DELEGATE,
       expirationDate: undefined,
     });
   });
 
-  test("passes options through", async ({ mockToken }) => {
-    const options = delegateDecryptionMutationOptions(mockToken, mockToken.address);
+  test("passes options through", async ({ sdk }) => {
+    const spy = vi.spyOn(sdk, "delegateDecryption").mockResolvedValue({ txHash: "0x" });
+    const options = delegateDecryptionMutationOptions(sdk, TOKEN);
     const expirationDate = new Date("2030-01-01");
 
-    await options.mutationFn({
-      delegateAddress: "0x2b2B2B2b2B2b2B2b2B2b2b2b2B2B2b2b2B2b2B2B",
-      expirationDate,
-    });
-    expect(mockToken.delegateDecryption).toHaveBeenCalledWith({
-      delegateAddress: "0x2b2B2B2b2B2b2B2b2B2b2b2b2B2B2b2b2B2b2B2B",
+    await options.mutationFn({ delegateAddress: DELEGATE, expirationDate });
+    expect(spy).toHaveBeenCalledWith({
+      contractAddress: TOKEN,
+      delegateAddress: DELEGATE,
       expirationDate,
     });
   });
