+++ packages/sdk/src/query/__tests__/resume-unshield.test.ts
@@ -2,21 +2,21 @@ import { describe, expect, test } from "../../test-fixtures";
 import { resumeUnshieldMutationOptions } from "../resume-unshield";
 
 describe("resumeUnshieldMutationOptions", () => {
-  test("delegates resumeUnshield", async ({ mockToken }) => {
-    const options = resumeUnshieldMutationOptions(mockToken, mockToken.address);
+  test("delegates resumeUnshield", async ({ mockWrappedToken }) => {
+    const options = resumeUnshieldMutationOptions(mockWrappedToken);
 
-    expect(options.mutationKey).toEqual(["zama.resumeUnshield", mockToken.address]);
+    expect(options.mutationKey).toEqual(["zama.resumeUnshield", mockWrappedToken.address]);
     await options.mutationFn({
       unwrapTxHash: "0xaAaAaAaaAaAaAaaAaAAAAAAAAaaaAaAaAaaAaaAaaaaaaaaaaaaaaaaaaaaaaaaa",
     });
-    expect(mockToken.resumeUnshield).toHaveBeenCalledWith(
+    expect(mockWrappedToken.resumeUnshield).toHaveBeenCalledWith(
       "0xaAaAaAaaAaAaAaaAaAAAAAAAAaaaAaAaAaaAaaAaaaaaaaaaaaaaaaaaaaaaaaaa",
       {},
     );
   });
 
-  test("passes flat callbacks to resumeUnshield", async ({ mockToken }) => {
-    const options = resumeUnshieldMutationOptions(mockToken, mockToken.address);
+  test("passes flat callbacks to resumeUnshield", async ({ mockWrappedToken }) => {
+    const options = resumeUnshieldMutationOptions(mockWrappedToken);
     const onUnwrapSubmitted = () => {};
     const onFinalizing = () => {};
 
@@ -25,7 +25,7 @@ describe("resumeUnshieldMutationOptions", () => {
       onUnwrapSubmitted,
       onFinalizing,
     });
-    expect(mockToken.resumeUnshield).toHaveBeenCalledWith(
+    expect(mockWrappedToken.resumeUnshield).toHaveBeenCalledWith(
       "0xaAaAaAaaAaAaAaaAaAAAAAAAAaaaAaAaAaaAaaAaaaaaaaaaaaaaaaaaaaaaaaaa",
       { onUnwrapSubmitted, onFinalizing },
     );
