+++ packages/sdk/src/query/__tests__/unshield.test.ts
@@ -2,21 +2,21 @@ import { describe, expect, test } from "../../test-fixtures";
 import { unshieldMutationOptions } from "../unshield";
 
 describe("unshieldMutationOptions", () => {
-  test("delegates unshield", async ({ mockToken }) => {
-    const options = unshieldMutationOptions(mockToken, mockToken.address);
+  test("delegates unshield", async ({ mockWrappedToken }) => {
+    const options = unshieldMutationOptions(mockWrappedToken);
 
-    expect(options.mutationKey).toEqual(["zama.unshield", mockToken.address]);
+    expect(options.mutationKey).toEqual(["zama.unshield", mockWrappedToken.address]);
     await options.mutationFn({ amount: 11n });
-    expect(mockToken.unshield).toHaveBeenCalledWith(11n, {});
+    expect(mockWrappedToken.unshield).toHaveBeenCalledWith(11n, {});
   });
 
-  test("delegates callbacks when provided", async ({ mockToken }) => {
-    const options = unshieldMutationOptions(mockToken, mockToken.address);
+  test("delegates callbacks when provided", async ({ mockWrappedToken }) => {
+    const options = unshieldMutationOptions(mockWrappedToken);
     const onUnwrapSubmitted = () => undefined;
     const onFinalizing = () => undefined;
 
     await options.mutationFn({ amount: 12n, onUnwrapSubmitted, onFinalizing });
-    expect(mockToken.unshield).toHaveBeenCalledWith(
+    expect(mockWrappedToken.unshield).toHaveBeenCalledWith(
       12n,
       expect.objectContaining({ onUnwrapSubmitted, onFinalizing }),
     );
