+++ packages/sdk/src/query/__tests__/unwrap-all.test.ts
@@ -2,11 +2,11 @@ import { describe, expect, test } from "../../test-fixtures";
 import { unwrapAllMutationOptions } from "../unwrap-all";
 
 describe("unwrapAllMutationOptions", () => {
-  test("delegates unwrapAll", async ({ mockToken }) => {
-    const options = unwrapAllMutationOptions(mockToken, mockToken.address);
+  test("delegates unwrapAll", async ({ mockWrappedToken }) => {
+    const options = unwrapAllMutationOptions(mockWrappedToken);
 
-    expect(options.mutationKey).toEqual(["zama.unwrapAll", mockToken.address]);
+    expect(options.mutationKey).toEqual(["zama.unwrapAll", mockWrappedToken.address]);
     await options.mutationFn();
-    expect(mockToken.unwrapAll).toHaveBeenCalled();
+    expect(mockWrappedToken.unwrapAll).toHaveBeenCalled();
   });
 });
