+++ packages/sdk/src/query/revoke-delegation.ts
@@ -1,6 +1,6 @@
-import type { Token } from "../token/token";
 import type { Address } from "viem";
 import type { TransactionResult } from "../types";
+import type { ZamaSDK } from "../zama-sdk";
 import type { MutationFactoryOptions } from "./factory-types";
 
 /** Variables for {@link revokeDelegationMutationOptions}. */
@@ -9,14 +9,16 @@ export interface RevokeDelegationParams {
 }
 
 export function revokeDelegationMutationOptions(
-  token: Token,
+  sdk: ZamaSDK,
+  contractAddress: Address,
 ): MutationFactoryOptions<
   readonly ["zama.revokeDelegation", Address],
   RevokeDelegationParams,
   TransactionResult
 > {
   return {
-    mutationKey: ["zama.revokeDelegation", token.address] as const,
-    mutationFn: async ({ delegateAddress }) => token.revokeDelegation({ delegateAddress }),
+    mutationKey: ["zama.revokeDelegation", contractAddress] as const,
+    mutationFn: async ({ delegateAddress }) =>
+      sdk.revokeDelegation({ contractAddress, delegateAddress }),
   };
 }
