+++ packages/sdk/src/query/shield.ts
@@ -1,7 +1,7 @@
-import type { Token } from "../token/token";
+import type { Address } from "viem";
+import type { WrappedToken } from "../token/wrapped-token";
 import type { ShieldOptions, TransactionResult } from "../types";
 import type { MutationFactoryOptions } from "./factory-types";
-import type { Address } from "viem";
 
 /** Variables for {@link shieldMutationOptions}. */
 export interface ShieldParams extends ShieldOptions {
@@ -9,7 +9,7 @@ export interface ShieldParams extends ShieldOptions {
 }
 
 export function shieldMutationOptions(
-  token: Token,
+  token: WrappedToken,
 ): MutationFactoryOptions<readonly ["zama.shield", Address], ShieldParams, TransactionResult> {
   return {
     mutationKey: ["zama.shield", token.address] as const,
