+++ packages/sdk/src/query/unwrap-all.ts
@@ -1,10 +1,10 @@
-import type { Token } from "../token/token";
+import type { WrappedToken } from "../token/wrapped-token";
 import type { TransactionResult } from "../types";
 import type { MutationFactoryOptions } from "./factory-types";
 import type { Address } from "viem";
 
 export function unwrapAllMutationOptions(
-  token: Token,
+  token: WrappedToken,
 ): MutationFactoryOptions<readonly ["zama.unwrapAll", Address], void, TransactionResult> {
   return {
     mutationKey: ["zama.unwrapAll", token.address] as const,
