+++ packages/sdk/src/token/__tests__/address-normalization.test.ts
@@ -1,32 +1,29 @@
 import { describe, expect, it } from "../../test-fixtures";
 
 import { Token } from "../token";
+import { WrappedToken } from "../wrapped-token";
 import type { Address } from "viem";
 
 describe("Address normalization (P6)", () => {
-  it("checksummed token address in constructor", ({ sdk }) => {
+  it("checksums Token address in constructor", ({ sdk }) => {
     const token = new Token(sdk, "0xABCDEF1234567890ABCDEF1234567890ABCDEF12" as Address);
 
     expect(token.address).toBe("0xabCDEF1234567890ABcDEF1234567890aBCDeF12");
   });
 
-  it("checksummed wrapper address in constructor", ({ sdk, tokenAddress }) => {
-    const token = new Token(
-      sdk,
-      tokenAddress,
-      "0xABCDEF1234567890ABCDEF1234567890ABCDEF12" as Address,
-    );
+  it("checksums WrappedToken address in constructor", ({ sdk }) => {
+    const token = new WrappedToken(sdk, "0xABCDEF1234567890ABCDEF1234567890ABCDEF12" as Address);
 
-    expect(token.wrapper).toBe("0xabCDEF1234567890ABcDEF1234567890aBCDeF12");
+    expect(token.address).toBe("0xabCDEF1234567890ABcDEF1234567890aBCDeF12");
   });
 
-  it("defaults wrapper to checksummed address when not provided", ({ sdk }) => {
-    const token = new Token(sdk, "0xABCDEF1234567890ABCDEF1234567890ABCDEF12" as Address);
-
-    expect(token.wrapper).toBe("0xabCDEF1234567890ABcDEF1234567890aBCDeF12");
-  });
-
-  it("rejects invalid address in constructor", ({ sdk }) => {
+  it("rejects invalid address in Token constructor", ({ sdk }) => {
     expect(() => new Token(sdk, "0xinvalid" as Address)).toThrow('Address "0xinvalid" is invalid.');
+  });
+
+  it("rejects invalid address in WrappedToken constructor", ({ sdk }) => {
+    expect(() => new WrappedToken(sdk, "0xinvalid" as Address)).toThrow(
+      'Address "0xinvalid" is invalid.',
+    );
   });
 });
