+++ packages/sdk/src/token/__tests__/shield.test.ts
@@ -5,10 +5,13 @@ import { ZamaSDKEvents } from "../../events/sdk-events";
 const UNDERLYING = "0x9C9c9c9c9c9c9C9c9c9C9C9c9c9C9c9c9c9c9C9c" as Address;
 const OTHER_RECIPIENT = "0x8b8b8b8b8B8B8b8B8B8b8b8b8b8B8B8B8B8b8B8b" as Address;
 
-describe("Token.shield", () => {
+describe("WrappedToken.shield", () => {
   // --- Callbacks (approveAndWrap path) ---
 
-  it("fires onApprovalSubmitted and onShieldSubmitted callbacks", async ({ token, provider }) => {
+  it("fires onApprovalSubmitted and onShieldSubmitted callbacks", async ({
+    wrappedToken: token,
+    provider,
+  }) => {
     vi.mocked(provider.readContract)
       .mockResolvedValueOnce(UNDERLYING) // underlying()
       .mockResolvedValueOnce(false) // supportsInterface (ERC-1363)
@@ -24,7 +27,10 @@ describe("Token.shield", () => {
     expect(onShieldSubmitted).toHaveBeenCalledWith("0xtxhash");
   });
 
-  it("skips onApprovalSubmitted when allowance is sufficient", async ({ token, provider }) => {
+  it("skips onApprovalSubmitted when allowance is sufficient", async ({
+    wrappedToken: token,
+    provider,
+  }) => {
     vi.mocked(provider.readContract)
       .mockResolvedValueOnce(UNDERLYING)
       .mockResolvedValueOnce(false) // supportsInterface (ERC-1363)
@@ -40,7 +46,7 @@ describe("Token.shield", () => {
     expect(onShieldSubmitted).toHaveBeenCalledOnce();
   });
 
-  it("completes shield even when callbacks throw", async ({ token, provider }) => {
+  it("completes shield even when callbacks throw", async ({ wrappedToken: token, provider }) => {
     vi.mocked(provider.readContract)
       .mockResolvedValueOnce(UNDERLYING)
       .mockResolvedValueOnce(false) // supportsInterface (ERC-1363)
@@ -59,7 +65,11 @@ describe("Token.shield", () => {
     expect(result.txHash).toBe("0xtxhash");
   });
 
-  it("passes to parameter for shield recipient", async ({ token, signer, provider }) => {
+  it("passes to parameter for shield recipient", async ({
+    wrappedToken: token,
+    signer,
+    provider,
+  }) => {
     vi.mocked(provider.readContract)
       .mockResolvedValueOnce(UNDERLYING)
       .mockResolvedValueOnce(false) // supportsInterface (ERC-1363)
@@ -73,7 +83,7 @@ describe("Token.shield", () => {
   });
 
   it("awaits approval receipt before submitting the wrap TX", async ({
-    token,
+    wrappedToken: token,
     signer,
     provider,
   }) => {
@@ -100,7 +110,7 @@ describe("Token.shield", () => {
   });
 
   it("performs full shield flow with exact approval", async ({
-    token,
+    wrappedToken: token,
     signer,
     relayer,
     handle: fixtureHandle,
@@ -146,8 +156,7 @@ describe("Token.shield", () => {
   // --- ERC-1363 routing ---
 
   describe("ERC-1363 routing", () => {
-    it("auto + ERC-1363 supported: uses transferAndCall on the underlying token", async ({
-      token,
+    it("auto + ERC-1363 supported: uses transferAndCall on the underlying token", async ({ wrappedToken: token,
       signer,
       provider,
     }) => {
@@ -165,8 +174,7 @@ describe("Token.shield", () => {
       );
     });
 
-    it("auto + ERC-1363 not supported: falls back to approve+wrap", async ({
-      token,
+    it("auto + ERC-1363 not supported: falls back to approve+wrap", async ({ wrappedToken: token,
       signer,
       provider,
     }) => {
@@ -190,8 +198,7 @@ describe("Token.shield", () => {
       );
     });
 
-    it("auto + supportsInterface reverts: falls back to approve+wrap", async ({
-      token,
+    it("auto + supportsInterface reverts: falls back to approve+wrap", async ({ wrappedToken: token,
       signer,
       provider,
     }) => {
@@ -211,8 +218,7 @@ describe("Token.shield", () => {
     // advertises ERC-1363 support but reverts on transferAndCall at runtime,
     // we throw rather than falling back. Users can opt out with
     // `shieldStrategy: "approveAndWrap"`.
-    it("auto + transferAndCall reverts at runtime: throws (no fallback)", async ({
-      token,
+    it("auto + transferAndCall reverts at runtime: throws (no fallback)", async ({ wrappedToken: token,
       signer,
       provider,
     }) => {
@@ -227,8 +233,7 @@ describe("Token.shield", () => {
       expect(signer.writeContract).toHaveBeenCalledOnce();
     });
 
-    it("auto + transferAndCall non-revert error (e.g. user rejection): does NOT fall back", async ({
-      token,
+    it("auto + transferAndCall non-revert error (e.g. user rejection): does NOT fall back", async ({ wrappedToken: token,
       signer,
       provider,
     }) => {
@@ -245,8 +250,7 @@ describe("Token.shield", () => {
       expect(signer.writeContract).toHaveBeenCalledOnce();
     });
 
-    it("auto + transferAndCall succeeds but receipt fails: does NOT fall back", async ({
-      token,
+    it("auto + transferAndCall succeeds but receipt fails: does NOT fall back", async ({ wrappedToken: token,
       signer,
       provider,
     }) => {
@@ -269,7 +273,7 @@ describe("Token.shield", () => {
   // --- Data encoding ---
 
   describe("transferAndCall data encoding", () => {
-    it("self-shield sends empty data (0x)", async ({ token, signer, provider }) => {
+    it("self-shield sends empty data (0x)", async ({ wrappedToken: token, signer, provider }) => {
       vi.mocked(provider.readContract)
         .mockResolvedValueOnce(UNDERLYING)
         .mockResolvedValueOnce(true)
@@ -286,8 +290,7 @@ describe("Token.shield", () => {
     // The wrapper decodes the recipient via address(bytes20(data)). We send the
     // raw 20-byte address (not ABI-encoded); ABI encoding would left-pad with
     // 12 zero bytes and bytes20() would slice them, corrupting the recipient.
-    it("shield-to-other sends raw 20-byte recipient address (not ABI-encoded)", async ({
-      token,
+    it("shield-to-other sends raw 20-byte recipient address (not ABI-encoded)", async ({ wrappedToken: token,
       signer,
       provider,
     }) => {
@@ -304,8 +307,7 @@ describe("Token.shield", () => {
       expect(callArgs.args[2]).toBe(getAddress(OTHER_RECIPIENT));
     });
 
-    it("explicit to=userAddress (case-insensitive) is treated as self-shield (data: 0x)", async ({
-      token,
+    it("explicit to=userAddress (case-insensitive) is treated as self-shield (data: 0x)", async ({ wrappedToken: token,
       signer,
       userAddress,
       provider,
@@ -329,8 +331,7 @@ describe("Token.shield", () => {
   // --- approvalStrategy interaction ---
 
   describe("approvalStrategy interaction with transferAndCall", () => {
-    it("approvalStrategy is ignored when transferAndCall path is used", async ({
-      token,
+    it("approvalStrategy is ignored when transferAndCall path is used", async ({ wrappedToken: token,
       signer,
       provider,
     }) => {
@@ -347,8 +348,7 @@ describe("Token.shield", () => {
       );
     });
 
-    it("onApprovalSubmitted callback is never fired on transferAndCall path", async ({
-      token,
+    it("onApprovalSubmitted callback is never fired on transferAndCall path", async ({ wrappedToken: token,
       provider,
     }) => {
       vi.mocked(provider.readContract)
@@ -366,7 +366,7 @@ describe("Token.shield", () => {
   // --- Detection caching ---
 
   describe("ERC-1363 detection caching", () => {
-    it("caches detection result across shield calls", async ({ token, provider }) => {
+    it("caches detection result across shield calls", async ({ wrappedToken: token, provider }) => {
       vi.mocked(provider.readContract)
         .mockResolvedValueOnce(UNDERLYING)
         .mockResolvedValueOnce(true)
@@ -380,7 +380,7 @@ describe("Token.shield", () => {
       expect(provider.readContract).toHaveBeenCalledTimes(4);
     });
 
-    it("isPayable() returns and caches detection result", async ({ token, provider }) => {
+    it("isPayable() returns and caches detection result", async ({ wrappedToken: token, provider }) => {
       vi.mocked(provider.readContract)
         .mockResolvedValueOnce(UNDERLYING)
         .mockResolvedValueOnce(true);
@@ -390,7 +390,7 @@ describe("Token.shield", () => {
       expect(provider.readContract).toHaveBeenCalledTimes(2);
     });
 
-    it("isPayable() caches false when supportsInterface reverts", async ({ token, provider }) => {
+    it("isPayable() caches false when supportsInterface reverts", async ({ wrappedToken: token, provider }) => {
       vi.mocked(provider.readContract)
         .mockResolvedValueOnce(UNDERLYING)
         .mockRejectedValueOnce(new Error("supportsInterface reverted"));
@@ -402,7 +402,7 @@ describe("Token.shield", () => {
       expect(provider.readContract).toHaveBeenCalledTimes(2);
     });
 
-    it("isPayable() caches false when underlying() reverts", async ({ token, provider }) => {
+    it("isPayable() caches false when underlying() reverts", async ({ wrappedToken: token, provider }) => {
       vi.mocked(provider.readContract).mockRejectedValueOnce(new Error("underlying() reverted"));
 
       expect(await token.isPayable()).toBe(false);
@@ -556,6 +556,6 @@ describe("Token.shield", () => {
     const options = shieldMutationOptions(mockToken);
 
     await options.mutationFn({ amount: 1n, approvalStrategy: "max" });
-    expect(mockToken.shield).toHaveBeenCalledWith(1n, { approvalStrategy: "max" });
+    expect(mockWrappedToken.shield).toHaveBeenCalledWith(1n, { approvalStrategy: "max" });
   });
 });
